add support for nova ssh user@host
the nova ssh command is convenient, but it is lacking some basic
niceties that everyone is used to with ssh, for instance actually
doing user@host to connect to an environment. This adds support
for "nova ssh user@host" to work as one would expect.

The functionality is added to both v1.1 and v3 clients. Tests for
this, and other ssh behavior are added to the v1.1 tree (the v3
test_shell.py has not synced over yet).

Change-Id: Ic4081f85c848507ebdc5e228ac345faf19127168